Add phobos and deimos, the martian moons. Change the lengths of the connections between planets by a scaling factor. Make the lengths of space connections between Mercury and Jupiter realistic, before the scaling by the above feature.
Add Phobos and Deimos, as well as the new asteroids in the kuiper belt introduced in 1.0.16/
 
New lengths for space connections may be slightly inaccurate due to the cutting off of decimal points and inaccuracies in the distances from the sun as well as the calculation.

When creating technologies for planets which don't have icons yet, the game crashes. This will fix that problem (by using the empty image when the Boolean value is true)

Add asteroids to the planets. They shouldn't be as frequent as done here, but we aren't even using our own definitions for space connections, so...